🦉 using Float64

+++ src/Model/src/grid/grids.jl
@@ -1,39 +1,39 @@
-const maxN = floor(Int, 1 / sqrt(eps(Float32)))
+const maxN = floor(Int, 1 / sqrt(eps(Float64)))
 
-Domain = Tuple{Float32, Float32}
+Domain = Tuple{Float64, Float64}
 
-struct Point <: FieldVector{3, Float32}
-    T::Float32
-    m::Float32
-    y::Float32
+struct Point <: FieldVector{3, Float64}
+    T::Float64
+    m::Float64
+    y::Float64
 end
 
-struct Policy <: FieldVector{2, Float32}
-    χ::Float32
-    α::Float32
+struct Policy <: FieldVector{2, Float64}
+    χ::Float64
+    α::Float64
 end
 
-struct Drift <: FieldVector{3, Float32}
-    dT::Float32
-    dm::Float32
-    dy::Float32
+struct Drift <: FieldVector{3, Float64}
+    dT::Float64
+    dm::Float64
+    dy::Float64
 end
 
-struct TransitionProbability <: FieldVector{6, Float32}
-    Δ₊T::Float32
-    Δ₊m::Float32
-    Δ₊y::Float32
-    Δ₋T::Float32
-    Δ₋m::Float32
-    Δ₋y::Float32
+struct TransitionProbability <: FieldVector{6, Float64}
+    Δ₊T::Float64
+    Δ₊m::Float64
+    Δ₊y::Float64
+    Δ₋T::Float64
+    Δ₋m::Float64
+    Δ₋y::Float64
 end
 
 PΔ₀(p::TransitionProbability) = 1 - sum(p)
 
 struct RegularGrid{N}
     X::Array{Point, 3}
-    h::Float32
-    Δ::NTuple{3, Float32}
+    h::Float64
+    Δ::NTuple{3, Float64}
     domains::AbstractArray{Domain}
 
     function RegularGrid(domains::AbstractVector{Domain}, N::Int)
@@ -43,7 +43,7 @@ struct RegularGrid{N}
 
         if N > maxN @warn "h < ϵ: ensure N > $maxN" end
 
-        h = Float32(inv(N))
+        h = Float64(inv(N))
         Ω = (range(d[1], d[2]; length = N) for d in domains) 
         Δ = ntuple(i -> domains[i][2] - domains[i][1], 3)
         
@@ -56,5 +56,5 @@ end
 
 Base.CartesianIndices(grid::RegularGrid) = CartesianIndices(grid.X)
 
-emptyscalarfield(grid::RegularGrid) = Array{Float32}(undef, size(grid))
-emptyvectorfield(grid::RegularGrid) = Array{Float32}(undef, size(grid)..., dimensions(grid))
+emptyscalarfield(grid::RegularGrid) = Array{Float64}(undef, size(grid))
+emptyvectorfield(grid::RegularGrid) = Array{Float64}(undef, size(grid)..., dimensions(grid))
