+++ src/Model/src/models/calibration.jl
@@ -1,5 +1,5 @@
 struct Calibration
     years::Vector{Int} # Years of the ipcc data
-    emissions::Vector{Float32} # Emissions in gton / year
-    γparameters::NTuple{3, Float32} # Paramters for γ
+    emissions::Vector{Float64} # Emissions in gton / year
+    γparameters::NTuple{3, Float64} # Paramters for γ
 end
