+++ src/Model/src/routines/bisection.jl
@@ -1,6 +1,6 @@
 @inline bisect(a, b) = a/2 + b/2
 
-function bisection(f, a_, b_; atol = 2eps(Float32), increasing = sign(f(b_)))
+function bisection(f, a_, b_; atol = 2eps(Float64), increasing = sign(f(b_)))
     a_, b_ = minmax(a_, b_)
     c = middle(a_, b_)
     z = f(c) * increasing
